minor refactoring to change class names to be more clear
